TIMER0: Correcting definition of hardware registers from pointer to u32 to u8

+++ lib/TIMER0_private.h
@@ -1,8 +1,8 @@
 #ifndef _TIMER0_PRIVATE_H
 #define _TIMER0_PRIVATE_H
 
-#define TCCR0 *((u32*) 0x53)
-#define TCNT0 *((u32*) 0x52)
-#define OCR0 *((u32*) 0x5C)
+#define TCCR0 *((u8*) 0x53)
+#define TCNT0 *((u8*) 0x52)
+#define OCR0 *((u8*) 0x5C)
 
 #endif
